Move models and controllers to separate folders
