+++ server/controllers/index.js
@@ -1,0 +1,2 @@
+module.exports.friends = require('./friends');
+module.exports.preferences = require('./preferences');
