+++ server/index.js
@@ -1,7 +1,7 @@
 require('dotenv').config();
 const path = require('path');
 const express = require('express');
-const controllers = require('./controllers');
+const { friends, preferences } = require("./controllers");
 const middleware = require('./middleware');
 
 const app = express();
@@ -11,10 +11,10 @@ app.use(express.static('client/public'));
 app.use(express.json());
 app.use(middleware.logger);
 
-app.get('/api/friends', controllers.getFriends);
+app.get('/api/friends', friends.getFriends);
 
-app.get('/api/survey', controllers.getSurvey);
-app.post('/api/users/:id/preferences', controllers.postUserPreferences);
+app.get('/api/survey', preferences.getSurvey);
+app.post('/api/users/:id/preferences', preferences.postUserPreferences);
 
 app.get('*', (req,res) =>{
   res.sendFile(path.join(__dirname + '../..' + '/client/public/index.html'));
